+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Buzzle".
 - Tests for "How time flies".
 - Tests for "Minesweeper level generator".
+- Tests for "In stereo".
 
 ### Changed
 - Renaming "Linear Bézier curves" to "Cubic Bézier curves".